+++ okapi-core/src/test/java/okapi/ModuleTest.java
@@ -46,6 +46,7 @@ import io.restassured.response.ValidatableResponse;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.json.Json;
 import org.folio.okapi.common.OkapiLogger;
 
 @java.lang.SuppressWarnings({"squid:S1192"})
@@ -354,7 +355,6 @@ public class ModuleTest {
    */
   @Test
   public void testFilters(TestContext context) {
-    //async = context.async();
     RestAssuredClient c;
     Response r;
 
@@ -452,7 +452,10 @@ public class ModuleTest {
       + "    \"methods\" : [ \"*\" ]," + LS
       + "    \"path\" : \"/\"," + LS
       + "    \"phase\" : \"PHASE\"," + LS // This will get replaced later
-      + "    \"type\" : \"request-response\"" + LS
+      + "    \"type\" : \"request-only\"" + LS
+      //      + "    \"type\" : \"request-response\"" + LS
+      // The only known use case for these uses req-only, so that's what we
+      // test with. Tested req-resp manually, and it seems to work too
       + "  } ]," + LS
       + "  \"requires\" : [ ]," + LS
       + "  \"launchDescriptor\" : {" + LS
@@ -478,20 +481,50 @@ public class ModuleTest {
     String locPostEnable = enableModule("post-f-module-1");
     logger.debug("testFilters post: " + locPostModule + " " + locationPostDeployment + " " + locPostEnable);
 
-    // Make a simple request. All three filters shold be called
+    // Make a simple GET request. All three filters shold be called
+    //
     c = api.createRestAssured3();
-    c.given()
+    traces = c.given()
       .header("X-Okapi-Tenant", okapiTenant)
       .header("X-Okapi-Token", okapiToken)
       .header("X-all-headers", "BL") // ask sample to report all headers
+      .header("X-filter-pre", "202") // ask pre-filter to return 202
+      .header("X-filter-post", "203") // ask post-filter to return 203
       .get("/testb")
       .then().statusCode(200)
       .log().ifValidationFails()
       .body(containsString("It works"))
       .extract().headers().getValues("X-Okapi-Trace");
+    logger.debug("Filter test. Traces: " + Json.encode(traces));
     Assert.assertTrue(traces.get(0).contains("GET auth-f-module-1"));
-    Assert.assertTrue(traces.get(1).contains("GET sample-f-module-1"));
-    logger.debug("testFilters made the last real call");
+    Assert.assertTrue(traces.get(1).contains("GET pre-f-module-1"));
+    Assert.assertTrue(traces.get(2).contains("GET sample-f-module-1"));
+    Assert.assertTrue(traces.get(3).contains("GET post-f-module-1"));
+
+    // Make a simple POST request. All three filters should be called
+    // test-module will return 200, which should not be
+    // overwritten by the pre and post-filters that returns 202 and 203
+    c = api.createRestAssured3();
+    traces = c.given()
+      .header("X-Okapi-Tenant", okapiTenant)
+      .header("X-Okapi-Token", okapiToken)
+      .header("X-all-headers", "BL") // ask sample to report all headers
+      .header("X-filter-pre", "202") // ask pre-filter to return 202
+      .header("X-filter-post", "203") // ask post-filter to return 203
+      // Those returns coders should be overwritten by the 200 from the handler
+      .body("Testing... ")
+      .post("/testb")
+      .then().statusCode(200)
+      .log().all() //ifValidationFails()
+      .body(containsString("Hello"))
+      .extract().headers().getValues("X-Okapi-Trace");
+    logger.debug("Filter test. Traces: " + Json.encode(traces));
+    Assert.assertTrue(traces.get(0).contains("POST auth-f-module-1"));
+    Assert.assertTrue(traces.get(1).contains("POST pre-f-module-1"));
+    Assert.assertTrue(traces.get(1).contains("202"));
+    Assert.assertTrue(traces.get(2).contains("POST sample-f-module-1"));
+    Assert.assertTrue(traces.get(3).contains("POST post-f-module-1"));
+    Assert.assertTrue(traces.get(3).contains("203"));
 
     // Clean up (in reverse order)
     logger.debug("testFilters starting to clean up");
@@ -1634,7 +1667,7 @@ public class ModuleTest {
       + "  }, {" + LS
       + "    \"id\" : \"_tenant\"," + LS
       + "    \"version\" : \"1.0\"" + LS // TODO - Define paths - add test
-      + "  } ]," + LS
+      + "  }]," + LS
       + "  \"launchDescriptor\" : {" + LS
       + "    \"exec\" : \"/usr/bin/false\"" + LS
       + "  }" + LS
@@ -1764,6 +1797,9 @@ public class ModuleTest {
       .then().statusCode(400);
 
     // Enable the sample
+    // Note that we can do this without the auth token. The test-auth module
+    // will create a non-login token certifying that we do not have a login,
+    // but will allow requests to any /_/ path,
     final String docEnableSample = "{" + LS
       + "  \"id\" : \"sample-module-1\"" + LS
       + "}";
@@ -1771,7 +1807,8 @@ public class ModuleTest {
     c.given()
       .header("Content-Type", "application/json")
       .body(docEnableSample).post("/_/proxy/tenants/" + okapiTenant + "/modules")
-      .then().statusCode(201)
+      .then().log().ifValidationFails()
+      .statusCode(201)
       .body(equalTo(docEnableSample));
     Assert.assertTrue("raml: " + c.getLastReport().toString(),
       c.getLastReport().isEmpty());
@@ -1838,18 +1875,16 @@ public class ModuleTest {
       .body(equalTo("No suitable module found for path /something.we.do.not.have"));
 
     // Request without an auth token
-    // This is acceptable, we get back a token that certifies that we have no
-    // logged-in username. We can use this for modulePermissions still.
-    // A real auth module would refuse the request because we do not have the
-    // permission. But the test-auth lets it pass...
+    // In theory, this is acceptable, we should get back a token that certifies
+    // that we have no logged-in username. We can use this for modulePermissions
+    // still. A real auth module would be likely to refuse the request because
+    // we do not have the necessary ModulePermissions. The auth module refuses it too.
     given()
       .header("X-Okapi-Tenant", okapiTenant)
       .header("X-all-headers", "B") // ask sample to report all headers
       .get("/testb")
-      .then()
-      .statusCode(200)
-      .body(containsString("X-Okapi-Token")) // auth created a token
-      .body(containsString("X-Okapi-User-Id:?"));  // with no good userid
+      .then().log().ifValidationFails()
+      .statusCode(401);
 
 
     // Failed login
@@ -1935,7 +1970,7 @@ public class ModuleTest {
     // Check that we don't do prefix matching
     given().header("X-Okapi-Tenant", okapiTenant)
       .header("X-Okapi-Token", okapiToken)
-      .get("/testbXXX")
+      .get("/testbZZZ")
       .then().statusCode(404);
 
     // Check that parameters don't mess with the routing
@@ -2586,7 +2621,7 @@ public class ModuleTest {
     Response r;
 
     RestAssuredClient c;
-    
+
     c = api.createRestAssured3();
     c.given().delete(locationSampleModule1).then().statusCode(204);
     Assert.assertTrue("raml: " + c.getLastReport().toString(),