+++ okapi-test-auth-module/src/main/java/org/folio/okapi/auth/Auth.java
@@ -134,6 +134,13 @@ class Auth {
       return;
     }
     ctx.response().putHeader("X-Auth-Filter-Phase", phase);
+    // Hack to test return codes on various filter phases
+    phase = phase.split(" ")[0];
+    String pHeader = ctx.request().headers().get("X-Filter-" + phase);
+    logger.debug("filter: 'X-Filter-" + phase + "': " + pHeader);
+    if (pHeader != null) {
+      ctx.response().setStatusCode(Integer.parseInt(pHeader));
+    }
     echo(ctx);
   }
 
@@ -147,8 +154,19 @@ class Auth {
     String userId = "?";
     String tok = ctx.request().getHeader(XOkapiHeaders.TOKEN);
     if (tok == null || tok.isEmpty()) {
-      logger.warn("test-auth: check called without " + XOkapiHeaders.TOKEN);
-      tok = token(tenant, "-"); // create a dummy token without username
+      if (!ctx.request().path().startsWith("/_/")) {
+        logger.warn("test-auth: check called without " + XOkapiHeaders.TOKEN
+          + " for " + ctx.request().path());
+        responseText(ctx, 401)
+          .end("test-auth: check called without " + XOkapiHeaders.TOKEN);
+        return;
+      } else {
+        tok = token(tenant, "-"); // create a dummy token without username
+        // We call /_/tenant and /_/tenantPermissions in our tests without a token.
+        // In real life, this is more complex, mod-authtoken creates a non-
+        // login token, possibly with modulePermissions, and then checks that
+        // against the permissions required for the tenant interface...
+      }
     } else {
       logger.debug("test-auth: check starting with tok " + tok + " and tenant " + tenant);